Unify hal
