+++ hal_interfaces/hal_interfaces/specific/amazon_warehouse/platform_controller.py
@@ -6,6 +6,39 @@ from std_msgs.msg import String
 
 DEFAULT = 2
 
+def cmdLift2String(cmdLift):
+    return cmdLift.msg
+
+class CMDLift ():
+
+    def __init__(self):
+        self.msg = String()
+        self.msg.data = "unload"
+    
+    def cmd(self,cmd):
+        self.msg.data = cmd
+
+    def __str__(self):
+        return "CMDlift:" + self.msg.data
+
+
+### HAL INTERFACE ###
+class PublisherPlatformNode(Node):
+
+    def __init__(self, topic):
+
+        super().__init__("PublisherPlatform")
+        self.pub = self.create_publisher(String, topic, 10)
+        self.data = CMDLift()
+
+    def load(self):
+        self.data.cmd("load")
+        self.pub.publish(cmdLift2String(self.data))
+
+    def unload(self):
+        self.data.cmd("unload")
+        self.pub.publish(cmdLift2String(self.data))
+
 class PlatformController(Node):
     def __init__(self, joint):
         '''
@@ -43,7 +76,6 @@ class PlatformController(Node):
         #print("self.future: ",self.future.result().success, self.future.result().status_message)
         self.future_lock.release()
 
-### HAL INTERFACE ###
 class PlatformCommandNode(Node):
 
     def __init__(self, topic):
@@ -78,4 +110,3 @@ class PlatformCommandNode(Node):
         
         self.controller.send_effort(float(effort))
         self.applied_effort = self.applied_effort + effort
-
